Move hyparquet.js to index.js
