+++ package.json
@@ -26,13 +26,13 @@
     "types"
   ],
   "type": "module",
-  "types": "types/hyparquet.d.ts",
-  "main": "src/hyparquet.js",
+  "types": "types/index.d.ts",
+  "main": "src/index.js",
   "exports": {
     ".": {
       "browser": {
-        "types": "./types/hyparquet.d.ts",
-        "import": "./src/hyparquet.js"
+        "types": "./types/index.d.ts",
+        "import": "./src/index.js"
       },
       "default": {
         "types": "./types/node.d.ts",
@@ -54,13 +54,13 @@
     "test": "vitest run"
   },
   "devDependencies": {
-    "@types/node": "22.15.21",
+    "@types/node": "22.15.29",
     "@vitest/coverage-v8": "3.1.4",
-    "eslint": "9.27.0",
+    "eslint": "9.28.0",
     "eslint-plugin-jsdoc": "50.6.17",
     "hyparquet-compressors": "1.1.1",
     "typescript": "5.8.3",
-    "typescript-eslint": "8.32.1",
+    "typescript-eslint": "8.33.0",
     "vitest": "3.1.4"
   }
 }
